헌찬 : chart 수정

+++ src/main/java/com/model/MonitoringVO.java
@@ -6,10 +6,10 @@ public class MonitoringVO {
 	private int monitoring_seq;
 	private int heartrate; // 심박수
 	private int o2; // 산소포화도
-	private int temp; // 체온
+	private double temp; // 체온
 	private String regdate; // 수집 날짜
 	
-	public MonitoringVO(String driver_id, int monitoring_seq, int heartrate, int o2, int temp, String regdate) {
+	public MonitoringVO(String driver_id, int monitoring_seq, int heartrate, int o2, double temp, String regdate) {
 		super();
 		this.driver_id = driver_id;
 		this.monitoring_seq = monitoring_seq;
@@ -34,7 +34,7 @@ public class MonitoringVO {
 	public int getO2() {
 		return o2;
 	}
-	public int getTemp() {
+	public double getTemp() {
 		return temp;
 	}
 	public String getRegdate() {
